remove controller-runtime dependency in the SDK
On-behalf-of: @SAP christoph.mewes@sap.com
