+++ sdk/apis/operator/v1alpha1/cacheserver_types.go
@@ -56,7 +56,3 @@ type CacheServerList struct {
 	metav1.ListMeta `json:"metadata,omitempty"`
 	Items           []CacheServer `json:"items"`
 }
-
-func init() {
-	SchemeBuilder.Register(&CacheServer{}, &CacheServerList{})
-}
